rag/requirements.txt
fastapi==0.115.5
groq==0.12.0
langchain==0.3.7
langchain_community==0.3.7
langchain_google_genai==2.0.4
langchain_text_splitters==0.3.2
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.32.0
tidb_vector
mysqlclient